refactor: Merge distributed execution into core ModularPipelines package
Move all 22 implementation files from ModularPipelines.Distributed into
ModularPipelines/Distributed/ subdirectories. Distributed infrastructure
(coordinator, artifact store, serialization) is now always registered with
in-memory defaults via DependencyInjectionSetup. When AddDistributedMode()
is called with TotalInstances > 1, PipelineBuilder conditionally replaces
the executor based on role — no plugin ceremony needed.

- Delete ModularPipelines.Distributed project entirely
- Remove plugin-based service registration (DistributedPipelinePlugin)
- Add RegisterDistributedServices() to DependencyInjectionSetup
- Add ActivateDistributedModeIfConfigured() to PipelineBuilder
- Simplify DistributedPipelineBuilderExtensions (no plugin registration)
- Remove Distributed ProjectReference from Redis, S3, and test projects
