+++ ModularPipelines.sln
@@ -125,8 +125,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModularPipelines.Syft", "sr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModularPipelines.Grype", "src\ModularPipelines.Grype\ModularPipelines.Grype.csproj", "{60E4E82D-7BBF-4513-80ED-36A2273BB97D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModularPipelines.Distributed", "src\ModularPipelines.Distributed\ModularPipelines.Distributed.csproj", "{FEE8EBAE-189B-4988-9C0D-12483838673A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModularPipelines.Distributed.UnitTests", "test\ModularPipelines.Distributed.UnitTests\ModularPipelines.Distributed.UnitTests.csproj", "{A5D634E1-4AE9-4EA6-AD4B-E7FE81F52749}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModularPipelines.Distributed.Redis", "src\ModularPipelines.Distributed.Redis\ModularPipelines.Distributed.Redis.csproj", "{C6028374-2C99-4770-A8DE-58DEAD1DE854}"
@@ -819,18 +817,6 @@ Global
 		{60E4E82D-7BBF-4513-80ED-36A2273BB97D}.Release|x64.Build.0 = Release|Any CPU
 		{60E4E82D-7BBF-4513-80ED-36A2273BB97D}.Release|x86.ActiveCfg = Release|Any CPU
 		{60E4E82D-7BBF-4513-80ED-36A2273BB97D}.Release|x86.Build.0 = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|x64.Build.0 = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Debug|x86.Build.0 = Debug|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|x64.ActiveCfg = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|x64.Build.0 = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|x86.ActiveCfg = Release|Any CPU
-		{FEE8EBAE-189B-4988-9C0D-12483838673A}.Release|x86.Build.0 = Release|Any CPU
 		{A5D634E1-4AE9-4EA6-AD4B-E7FE81F52749}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A5D634E1-4AE9-4EA6-AD4B-E7FE81F52749}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A5D634E1-4AE9-4EA6-AD4B-E7FE81F52749}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -952,7 +938,6 @@ Global
 		{0FB125FE-5AB3-4667-8D1B-85A6284474ED} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{2E70AA19-0309-4C6F-83D2-8E3DD2A7EC89} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{60E4E82D-7BBF-4513-80ED-36A2273BB97D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
-		{FEE8EBAE-189B-4988-9C0D-12483838673A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{A5D634E1-4AE9-4EA6-AD4B-E7FE81F52749} = {F213898F-1E32-48F1-AB8C-83D2BD01A93B}
 		{C6028374-2C99-4770-A8DE-58DEAD1DE854} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{97523284-1AB2-4BB1-84A2-818E103C6DE7} = {F213898F-1E32-48F1-AB8C-83D2BD01A93B}
