+++ src/ModularPipelines.Build/ModularPipelines.Build.csproj
@@ -12,7 +12,6 @@
     <ProjectReference Include="..\ModularPipelines.Git\ModularPipelines.Git.csproj" />
     <ProjectReference Include="..\ModularPipelines.Node\ModularPipelines.Node.csproj" />
     <ProjectReference Include="..\ModularPipelines\ModularPipelines.csproj" />
-    <ProjectReference Include="..\ModularPipelines.Distributed\ModularPipelines.Distributed.csproj" />
     <ProjectReference Include="..\ModularPipelines.Distributed.Redis\ModularPipelines.Distributed.Redis.csproj" />
     <ProjectReference Include="..\ModularPipelines.Distributed.Artifacts.S3\ModularPipelines.Distributed.Artifacts.S3.csproj" />
   </ItemGroup>
