+++ src/ModularPipelines.Distributed.Artifacts.S3/ModularPipelines.Distributed.Artifacts.S3.csproj
@@ -7,7 +7,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ModularPipelines\ModularPipelines.csproj" />
-    <ProjectReference Include="..\ModularPipelines.Distributed\ModularPipelines.Distributed.csproj" />
   </ItemGroup>
 
   <ItemGroup>
