+++ src/ModularPipelines/Distributed/Artifacts/ArtifactContextImpl.cs
