+++ src/ModularPipelines/Distributed/Artifacts/ArtifactLifecycleManager.cs
