+++ src/ModularPipelines/Distributed/Artifacts/InMemoryDistributedArtifactStore.cs
