+++ src/ModularPipelines/Distributed/Capabilities/OsCapabilityDetector.cs
