+++ src/ModularPipelines/Distributed/Configuration/RoleDetector.cs
