+++ src/ModularPipelines/Distributed/Coordination/InMemoryDistributedCoordinator.cs
