+++ src/ModularPipelines/Distributed/Extensions/ArtifactContextExtensions.cs
