+++ src/ModularPipelines/Distributed/Master/DistributedResultCollector.cs
