+++ src/ModularPipelines/Distributed/Master/DistributedSummaryAggregator.cs
