+++ src/ModularPipelines/Distributed/Master/WorkerHealthMonitor.cs
