+++ src/ModularPipelines/Distributed/Serialization/ModuleResultSerializer.cs
