+++ src/ModularPipelines/Distributed/Serialization/ModuleTypeRegistry.cs
