+++ src/ModularPipelines/Distributed/Worker/WorkerCancellationMonitor.cs
