+++ src/ModularPipelines/ModularPipelines.csproj
@@ -36,11 +36,6 @@
       <_Parameter1>ModularPipelines.TeamCity</_Parameter1>
     </AssemblyAttribute>
 
-    <!-- Distributed package - accesses internal scheduler/executor interfaces -->
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>ModularPipelines.Distributed</_Parameter1>
-    </AssemblyAttribute>
-
     <!-- Test assemblies - verify internal behavior -->
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
       <_Parameter1>ModularPipelines.UnitTests</_Parameter1>
