+++ test/ModularPipelines.Distributed.Artifacts.S3.UnitTests/ModularPipelines.Distributed.Artifacts.S3.UnitTests.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\ModularPipelines\ModularPipelines.csproj" />
-    <ProjectReference Include="..\..\src\ModularPipelines.Distributed\ModularPipelines.Distributed.csproj" />
     <ProjectReference Include="..\..\src\ModularPipelines.Distributed.Artifacts.S3\ModularPipelines.Distributed.Artifacts.S3.csproj" />
   </ItemGroup>
