+++ test/ModularPipelines.Distributed.UnitTests/ModularPipelines.Distributed.UnitTests.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\ModularPipelines\ModularPipelines.csproj" />
-    <ProjectReference Include="..\..\src\ModularPipelines.Distributed\ModularPipelines.Distributed.csproj" />
   </ItemGroup>
 
 </Project>
